requirements.txt
Flask==1.0.2
krakenex==2.1.0
google-cloud-datastore==1.10.0
google-cloud-bigquery==1.22.0
google-cloud-logging==1.14.0
